Fixing eoshapl to properly digest SIGHUP and to prevent failover misconfiguration (target1=target2). Adding Report class to decode file transaction reports. Adding simple 'io stat' view for 1min, 5min & 1hour IO measurements. Returning error if 'vid rm' (vid remove gateway) fails because entry does not exist. Putten DrainJob in 'lostfiles' state if some files could not been drained. Adding graceperiod,drainperiod etc. as global variables to a space view. (later we will inherit them into the filesystems if we move between spaces or add new to a space). Writing eosha PID into /var/run/eosha.pid.

+++ mgm/Vid.cc
@@ -200,17 +200,24 @@ Vid::Rm(XrdOucEnv &env, int &retc, XrdOucString &stdOut, XrdOucString &stdErr)
     retc = EINVAL;
     return false;
   }
-  
-  stdOut += "success: rm vid [ "; stdOut += inenv; stdOut += "]";
-  errno = 0;
-  retc = 0;
 
-  eos::common::Mapping::gVirtualUidMap.erase(skey.c_str());
-  eos::common::Mapping::gVirtualGidMap.erase(skey.c_str());
+  int nerased=0;
+  nerased += eos::common::Mapping::gVirtualUidMap.erase(skey.c_str());
+  nerased += eos::common::Mapping::gVirtualGidMap.erase(skey.c_str());
 
-  gOFS->ConfEngine->DeleteConfigValue("vid",skey.c_str());
+  gOFS->ConfEngine->DeleteConfigValue("vid",skey.c_str());  
 
-  return true;
+  if (nerased) {
+    stdOut += "success: rm vid [ "; stdOut += inenv; stdOut += "]";
+    errno = 0;
+    retc = 0;
+    return true;
+  } else {
+    stdErr += "error: nothing has been removed";
+    errno = EINVAL;
+    retc = EINVAL;
+    return false;
+  }
 }
 
 /*----------------------------------------------------------------------------*/
